server: move settings tests to new package settings/integration_tests
Release note: None
