+++ pkg/BUILD.bazel
@@ -316,6 +316,7 @@ ALL_TESTS = [
     "//pkg/server/tenantsettingswatcher:tenantsettingswatcher_test",
     "//pkg/server/tracedumper:tracedumper_test",
     "//pkg/server:server_test",
+    "//pkg/settings/integration_tests:integration_tests_test",
     "//pkg/settings/lint:lint_test",
     "//pkg/settings/rulebasedscanner:rulebasedscanner_test",
     "//pkg/settings:settings_test",
@@ -1597,6 +1598,7 @@ GO_TARGETS = [
     "//pkg/server:server",
     "//pkg/server:server_test",
     "//pkg/settings/cluster:cluster",
+    "//pkg/settings/integration_tests:integration_tests_test",
     "//pkg/settings/lint:lint_test",
     "//pkg/settings/rulebasedscanner:rulebasedscanner",
     "//pkg/settings/rulebasedscanner:rulebasedscanner_test",
