+++ pkg/server/BUILD.bazel
@@ -455,7 +455,6 @@ go_test(
         "server_systemlog_gc_test.go",
         "server_test.go",
         "settings_cache_test.go",
-        "settings_test.go",
         "span_stats_server_test.go",
         "span_stats_test.go",
         "statements_test.go",
